requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==1.0.0
python-multipart==0.0.6
openai==1.3.5
jinja2==3.1.2
aiofiles==23.2.1
pyyaml==6.0.1
python-logging-loki==0.3.1
tenacity==8.2.3 